Use exact-width types
The code pretty much assumed long was 32 bits and long long was 64.
This caused problems when that was no longer the case (64-bit Linux).
Explicitly use 32-bit and 64-bit types.

+++ histogram.h
@@ -14,7 +14,7 @@
 typedef int H_int;
 
 typedef struct {
-	long  total_sum;
+	int32_t total_sum;
 	H_int total_count;
 	int   max_index;
 	H_int max_count;
